Fix line width anti-aliasing issues
* See #650
* Adjust figure bounds according to the line width
* AbstractDiagramModelObjectFigure#setLineWidth() only accepts integers
* Use ExtendedSWTGraphics class so that we can determine the scale when exporting to image

+++ com.archimatetool.editor/src/com/archimatetool/editor/diagram/figures/AbstractFigureDelegate.java
@@ -110,7 +110,7 @@ public class AbstractFigureDelegate implements IFigureDelegate {
      * @param lineWidth The line width
      * @param bounds The bounds of the object
      */
-    protected void setLineWidth(Graphics graphics, float lineWidth, Rectangle bounds) {
+    protected void setLineWidth(Graphics graphics, int lineWidth, Rectangle bounds) {
         if(getOwner() instanceof AbstractDiagramModelObjectFigure) {
             ((AbstractDiagramModelObjectFigure)getOwner()).setLineWidth(graphics, lineWidth, bounds);
         }
